separate control from sink
