+++ Makefile
@@ -8,10 +8,11 @@ LDFLAGS = -lzmq -lczmq -ljansson
 
 SYSTEMD_LDFLAGS = -lsystemd
 
-default: journal-gateway-zmtp-source journal-gateway-zmtp-sink
+default: journal-gateway-zmtp-source journal-gateway-zmtp-sink journal-gateway-zmtp-control
 
 source: journal-gateway-zmtp-source
 sink: journal-gateway-zmtp-sink
+control: journal-gateway-zmtp-control
 
 journal-gateway-zmtp-source: journal-gateway-zmtp-source.o
 	$(CC) journal-gateway-zmtp-source.o $(LDFLAGS) $(SYSTEMD_LDFLAGS) -o journal-gateway-zmtp-source
@@ -25,6 +26,12 @@ journal-gateway-zmtp-sink: journal-gateway-zmtp-sink.o
 journal-gateway-zmtp-sink.o: $(SRC_DIR)/journal-gateway-zmtp-sink.c
 	$(CC) $(CFLAGS) $(SRC_DIR)/journal-gateway-zmtp-sink.c -o journal-gateway-zmtp-sink.o
 
+journal-gateway-zmtp-control: journal-gateway-zmtp-control.o
+	$(CC) journal-gateway-zmtp-control.o $(LDFLAGS) -o journal-gateway-zmtp-control
+
+journal-gateway-zmtp-control.o:$(SRC_DIR)/journal-gateway-zmtp-control.c
+	$(CC) $(CFLAGS) $(SRC_DIR)/journal-gateway-zmtp-control.c -o journal-gateway-zmtp-control.o
+
 clean:
-	rm -f *.o journal-gateway-zmtp-source journal-gateway-zmtp-sink
+	rm -f *.o journal-gateway-zmtp-source journal-gateway-zmtp-sink journal-gateway-zmtp-control
